--- Demo/DemoSnapshotTests.swift
@@ -1,139 +1,0 @@
-//
-//  Copyright © 2018 FINN AS. All rights reserved.
-//
-
-import FBSnapshotTestCase
-import FinniversKit
-import Demo
-
-class DnaViewTests: FBSnapshotTestCase {
-    override func setUp() {
-        super.setUp()
-        recordMode = false
-    }
-
-    func testDnaViewsColor() {
-        snapshot(.color)
-    }
-
-    func testDnaViewsFont() {
-        snapshot(.font)
-    }
-
-    func testDnaViewsSpacing() {
-        snapshot(.spacing)
-    }
-
-    func testDnaViewsAssets() {
-        snapshot(.assets)
-    }
-}
-
-class ComponentViewTests: FBSnapshotTestCase {
-    override func setUp() {
-        super.setUp()
-        recordMode = false
-    }
-
-    func testBroadcast() {
-        snapshot(.broadcast)
-    }
-
-    func testBroadcastContainer() {
-        snapshot(.broadcastContainer)
-    }
-
-    func testButton() {
-        snapshot(.button)
-    }
-
-    func testLabel() {
-        snapshot(.label)
-    }
-
-    func testRibbon() {
-        snapshot(.ribbon)
-    }
-
-    func testTextField() {
-        snapshot(.textField)
-    }
-
-    func testToast() {
-        snapshot(.toast)
-    }
-
-    func testSwitchView() {
-        snapshot(.switchView)
-    }
-
-    func testInlineConsent() {
-        snapshot(.inlineConsent)
-    }
-
-    func testConsentTransparencyInfo() {
-        snapshot(.consentTransparencyInfo)
-    }
-
-    func testCheckbox() {
-        snapshot(.checkbox)
-    }
-
-    func testRadioButton() {
-        snapshot(.radioButton)
-    }
-
-    func testRoundedImageView() {
-        snapshot(.roundedImageView)
-    }
-}
-
-class RecyclingViewTests: FBSnapshotTestCase {
-    override func setUp() {
-        super.setUp()
-        recordMode = false
-    }
-
-    func testNotificationsListView() {
-        snapshot(.notificationsListView)
-    }
-
-    func testMarketsGridView() {
-        snapshot(.marketsGridView)
-    }
-
-    func testAdsGridView() {
-        snapshot(.adsGridView)
-    }
-
-    func testFavoritesListView() {
-        snapshot(.favoritesListView)
-    }
-}
-
-class FullscreenViewTests: FBSnapshotTestCase {
-    override func setUp() {
-        super.setUp()
-        recordMode = false
-    }
-
-    func testFrontpageView() {
-        snapshot(.frontpageView)
-    }
-
-    func testPopupView() {
-        snapshot(.popupView)
-    }
-
-    func testEmptyView() {
-        snapshot(.emptyView)
-    }
-
-    func testReportAdView() {
-        snapshot(.reportAdView)
-    }
-
-    func testReviewView() {
-        snapshot(.reviewView)
-    }
-}
